chore: Interface 파일 분리
